Proper fix gcc -Wsign-compare

+++ tinyxml2.cpp
@@ -1861,7 +1861,7 @@ XMLError XMLDocument::LoadFile( FILE* fp )
         return _errorID;
     }
 
-    if ( (size_t)filelength >= (size_t)-1 ) {
+    if ( (unsigned long)filelength >= (size_t)-1 ) {
         // Cannot handle files which won't fit in buffer together with null terminator
         SetError( XML_ERROR_FILE_READ_ERROR, 0, 0 );
         return _errorID;
